requirements.txt
python-telegram-bot==20.7
openai==1.54.3
google-generativeai==0.8.3
supabase==2.7.4
python-dotenv==1.0.0
loguru==0.7.2
aiohttp==3.10.10
Pillow==10.4.0